Should be a short.

+++ src/main/java/org/greatmancode/vanillaclassic/protocol/msg/MessageMessage.java
@@ -27,15 +27,15 @@
 package org.greatmancode.vanillaclassic.protocol.msg;
 
 public class MessageMessage extends VanillaClassicMessage {
-	private final byte playerID;
+	private final short playerID;
 	private final String message;
 
-	public MessageMessage(byte playerID, String message) {
+	public MessageMessage(short playerID, String message) {
 		this.playerID = playerID;
 		this.message = message;
 	}
 
-	public byte getPlayerID() {
+	public short getPlayerID() {
 		return playerID;
 	}
